Move matchers to matchers package and tidy up unused code
