+++ app/controllers/RulesController.scala
@@ -1,5 +1,6 @@
 package controllers
 
+import matchers.{LanguageToolFactory, RegexMatcher}
 import model.{Category, RegexRule}
 import play.api.mvc._
 import rules.SheetsRuleResource
