+++ app/matchers/LanguageToolMatcher.scala
@@ -1,17 +1,17 @@
-package services
+package matchers
 
 import java.io.File
 
-import org.languagetool._
-import org.languagetool.rules.{Rule => LanguageToolRule}
-import org.languagetool.rules.spelling.morfologik.suggestions_ordering.SuggestionsOrdererConfig
-
-import collection.JavaConverters._
-import scala.concurrent.{ExecutionContext, Future}
-import play.api.Logging
 import model.{LTRule, RuleMatch}
-import org.languagetool.rules.CategoryId
+import org.languagetool._
+import org.languagetool.rules.spelling.morfologik.suggestions_ordering.SuggestionsOrdererConfig
+import org.languagetool.rules.{CategoryId, Rule => LanguageToolRule}
+import play.api.Logging
+import services.MatcherRequest
 import utils.Matcher
+
+import scala.collection.JavaConverters._
+import scala.concurrent.{ExecutionContext, Future}
 
 class LanguageToolFactory(
                            maybeLanguageModelDir: Option[File],
@@ -47,11 +47,15 @@ class LanguageToolFactory(
     }
     instance.enableRuleCategory(new CategoryId(category))
 
-    (new LanguageTool(category, instance), ruleIngestionErrors)
+    (new LanguageToolMatcher(category, instance), ruleIngestionErrors)
   }
 }
 
-class LanguageTool(category: String, instance: JLanguageTool)(implicit ec: ExecutionContext) extends Matcher {
+
+/**
+  * A Matcher that wraps a LanguageTool instance.
+  */
+class LanguageToolMatcher(category: String, instance: JLanguageTool)(implicit ec: ExecutionContext) extends Matcher {
   def getId = "language-tool"
 
   def getCategory = category
